feat(customization): support customizing catalog and api index pages via outlets
