+++ .changeset/tender-ducks-tie.md
@@ -1,0 +1,6 @@
+---
+'@backstage/plugin-api-docs': patch
+'@backstage/plugin-catalog': patch
+---
+
+Support customizing index page layouts via outlets
