+++ plugins/api-docs/src/components/ApiExplorerPage/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2021 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,4 +14,6 @@
  * limitations under the License.
  */
 
-export { ApiExplorerPage } from './ApiExplorerPage';
+export { ApiExplorerPage as ApiExplorerIndexPage } from './ApiExplorerPage';
+export { DefaultApiExplorerPage } from './DefaultApiExplorerPage';
+export type { DefaultApiExplorerPageProps } from './DefaultApiExplorerPage';
