+++ plugins/api-docs/src/components/index.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+export * from './ApiExplorerPage';
 export * from './ApiDefinitionCard';
 export * from './ApisCards';
 export * from './AsyncApiDefinitionWidget';
