+++ plugins/catalog/src/components/CatalogPage/index.ts
@@ -14,3 +14,5 @@
  * limitations under the License.
  */
 export { CatalogPage } from './CatalogPage';
+export { DefaultCatalogPage } from './DefaultCatalogPage';
+export type { DefaultCatalogPageProps } from './DefaultCatalogPage';
