+++ plugins/catalog/src/index.ts
@@ -54,3 +54,4 @@ export {
 
 export type { EntityLinksEmptyStateClassKey } from './components/EntityLinksCard';
 export type { SystemDiagramCardClassKey } from './components/SystemDiagramCard';
+export type { DefaultCatalogPageProps } from './components/CatalogPage';
